Smarter mamp folder

+++ Rakefile
@@ -37,16 +37,17 @@ task :install do
   config_folder = "#{ENV["HOME"]}/.mamp"
   php = `php -i | grep 'Loaded Configuration File'`[/(?<=\=>\s).+$/]
   targets = [
-    '/private/var/log/apache2',
-    '/etc/hosts',
-    '/etc/apache2/extra/httpd-vhosts.conf',
-    '/etc/apache2/httpd.conf',
+    "/etc/hosts",
+    "/var/log/apache2",
+    "/private/var/log/apache2",
+    "/etc/apache2/extra/httpd-vhosts.conf",
+    "/etc/apache2/httpd.conf",
     php
   ]
 
   Dir.mkdir(config_folder) unless File.exists?(config_folder)
   targets.each do |target|
-    unless File.exists?(config_folder + target[/\/[^\/]+$/])
+    if !File.exists?(File.join(config_folder,  target[/[^\/]+$/])) && File.exists?(target)
       `ln -s "#{target}" "#{config_folder}/#{target[/[^\/]+$/]}"`
     end
   end
